Fixed issue with form in config.js
The issue was originally described as:
Weird behavior on desktop:
    1) change RR field to option #1
    2) change display app target field
    3) change RR field to option #2
    4) change RR field to option #1
    Problem: display app target field doesn't reset
    Bigger problem: outputField gets stuck

I fixed the problem where outputField was stuck.
The issue was that each of the selection components have a v-model variable called selected.
However, selected was never reset when the RelatedRecordsField value was changed.
So the other 3 dropdown menus would hold onto old selected values regardless.

I created a method in each of the components ending with "select" to clear the selected variable.
Then I called the methods in the "Computation" component whenever "handleRRSelection" is run.