Fixed earlier bug
Earlier the dropdown would be blank if the previusly saved selection for the dropdown no longer
existed in the list of dropdown entries that could be picked.
This issue is now fixed.

To do this I added a new property in the data object which is used by vue.
The property is called fullyLoaded, it is a boolean which represts whether
the api call to kintone to retrieve values such as relatedRecords, etc has compleated.

Now in the root vue component (vm) there is a v-if in the outer div which checks the fullyLoaded variable.
Until fullyLoaded is set to true, none of the other components in the root component's template are loaded.
This fixes a lot of issues such as relatedRecords being undefined or not having the actual saved values.

This is a stable version where everything (at least I think so) works.