Finished adding dropdown error handling feature
fixed bugs:
If a user changes the field codes or field labels of fields that
have calculations associated with them. When they go into the settings page
for this plugin, normally the dropdowns would have been blank. Now they say "Select a field".
Also next to the dropdown with the problem now, there will be red text that shows up and
displays their previous selection. This can be helpfull to debug.

- Dropdowns not blank but say "Select a field".
- Red text shows previous selection next to dropdown in case of error.